scraper/requirements.txt
pandas==2.0.2
psycopg2_binary==2.9.9
python-dotenv==1.0.1
selenium==4.18.1
tqdm==4.66.2
requests==2.31.0
